requirements.txt
absl-py==2.3.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
decorator==4.4.2
ecdsa==0.19.1
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.1
fsspec==2025.5.1
h11==0.16.0
httptools==0.6.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jax==0.6.1
jaxlib==0.6.1
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
moviepy==1.0.3
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.9.0.80
opt_einsum==3.4.0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
proglog==0.1.12
protobuf==4.25.8
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
scipy==1.15.3
sentencepiece==0.2.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.46.2
sympy==1.14.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.146
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wheel==0.45.1
